✏️ update : 프로필 사진 변경 코드 추가
프로필 사진 변경 코드 추가

> Related to : #22
