+++ app/src/main/java/com/shootit/greme/network/SignOutInterface.kt
@@ -7,6 +7,6 @@ import retrofit2.http.Path
 
 interface SignOutInterface {
     // 회원 탈퇴
-    @PATCH("/user/{userId}")
-    fun signOut(@Path("userId") userId: Int) : Call<Void>
+    @PATCH("/user")
+    fun signOut() : Call<Void>
 }
