--- app/src/main/java/com/shootit/greme/ui/fragment/ProfileEditFragment.kt
@@ -1,126 +1,0 @@
-package com.shootit.greme.ui.fragment
-
-import android.Manifest
-import android.app.Activity
-import android.content.Intent
-import android.content.pm.PackageManager
-import android.net.Uri
-import android.os.Build
-import android.os.Bundle
-import android.provider.MediaStore
-import androidx.fragment.app.Fragment
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import androidx.activity.result.contract.ActivityResultContracts
-import androidx.core.content.ContextCompat
-import com.bumptech.glide.Glide
-import com.bumptech.glide.request.RequestOptions
-import com.shootit.greme.R
-import com.shootit.greme.databinding.FragmentProfileEditBinding
-import java.io.File
-
-class ProfileEditFragment : Fragment() {
-    // 전역 변수로 바인딩 객체 선언
-    private var mBinding: FragmentProfileEditBinding? = null
-    // 매번 null 체크를 할 필요없이 편의성을 위해 바인딩 변수 재선언
-    private val binding get() = mBinding!!
-    var imageFile : File? = null
-
-    companion object {
-        const val REVIEW_MIN_LENGTH = 10
-        // 갤러리 권한 요청
-        const val REQ_GALLERY = 1
-        // API 호출시 Parameter Key 값
-        const val PARAM_KEY_IMAGE = "image"
-        const val PARAM_KEY_PRODUCT_ID = "product_id"
-        const val PARAM_KEY_REVIEW = "review_content"
-        const val PARAM_KEY_RATING = "rating"
-    }
-    // 이미지를 결과값으로 받는 변수
-    private val imageResult = registerForActivityResult(
-        ActivityResultContracts.StartActivityForResult()){
-            result ->
-        if (result.resultCode == Activity.RESULT_OK){
-            // 이미지를 받으면 ImageView에 적용
-            val imageUri = result.data?.data
-            imageUri?.let{
-                // 서버 업로드를 위해 파일 형태로 변환
-                imageFile = File(getRealPathFromURI(it))
-
-                // 이미지를 불러온다
-                Glide.with(this)
-                    .load(imageUri)
-                    .fitCenter()
-                    .apply(RequestOptions().override(500,500))
-                    .into(binding.ivProfileEdit)
-            }
-        }
-    }
-
-    override fun onCreateView(
-        inflater: LayoutInflater, container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        // 바인딩
-        mBinding = FragmentProfileEditBinding.inflate(inflater, container, false)
-        val root: View = binding.root
-        return root
-    }
-
-    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-        super.onViewCreated(view, savedInstanceState)
-        binding.btnProfilePlus.setOnClickListener {
-            selectGallery()
-        }
-        binding.ivBack.setOnClickListener {
-            val settingFragment = SettingFragment()
-            requireActivity().supportFragmentManager
-                .beginTransaction()
-                .replace(R.id.nav_fl, settingFragment)
-                .addToBackStack(null)
-                .commit()
-        }
-    }
-    // 이미지 실제 경로 반환
-    fun getRealPathFromURI(uri: Uri): String {
-        val buildName = Build.MANUFACTURER
-        if(buildName.equals("Xiaomi")){
-            return uri.path!!
-        }
-        var columnIndex = 0
-        val proj = arrayOf(MediaStore.Images.Media.DATA)
-        val cursor = requireActivity().contentResolver.query(uri, proj, null,null,null)
-        if(cursor!!.moveToFirst()){
-            columnIndex = cursor.getColumnIndexOrThrow(MediaStore.Images.Media.DATA)
-        }
-        val result = cursor.getString(columnIndex)
-        cursor.close()
-        return result
-    }
-    // 갤러리를 부르는 메서드
-    private fun selectGallery(){
-        val writePermission = ContextCompat.checkSelfPermission(requireContext(), Manifest.permission.WRITE_EXTERNAL_STORAGE)
-        val readPermission = ContextCompat.checkSelfPermission(requireContext(), Manifest.permission.READ_EXTERNAL_STORAGE)
-
-        // 권한 확인
-        if(writePermission == PackageManager.PERMISSION_DENIED || readPermission == PackageManager.PERMISSION_DENIED){
-            // 권한 요청
-            requestPermissions(arrayOf(Manifest.permission.WRITE_EXTERNAL_STORAGE, Manifest.permission.READ_EXTERNAL_STORAGE),
-                DiaryFragment.REQ_GALLERY
-            )
-        }else{
-            // 권한이 있는 경우 갤러리 실행
-            val intent = Intent(Intent.ACTION_PICK)
-            // intent의 data와 type을 동시에 설정하는 메서드
-            intent.setDataAndType(
-                MediaStore.Images.Media.EXTERNAL_CONTENT_URI,"image/*"
-            )
-            imageResult.launch(intent)
-        }
-    }
-    override fun onDestroyView() {
-        super.onDestroyView()
-        mBinding = null
-    }
-}
