+++ app/src/main/java/com/shootit/greme/ui/view/DiaryDetailActivity.kt
@@ -12,6 +12,8 @@ import com.shootit.greme.databinding.ActivityDiaryDetailBinding
 import com.shootit.greme.model.DiaryImgData
 import com.shootit.greme.model.ResponseDateDiaryData
 import com.shootit.greme.network.ConnectionObject
+import okhttp3.MediaType
+import okhttp3.RequestBody
 import retrofit2.Call
 import retrofit2.Callback
 import retrofit2.Response
