--- app/src/main/java/com/shootit/greme/ui/view/ProfileEditActivity.kt
@@ -1,12 +1,0 @@
-package com.shootit.greme.ui.view
-
-import androidx.appcompat.app.AppCompatActivity
-import android.os.Bundle
-import com.shootit.greme.R
-
-class ProfileEditActivity : AppCompatActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_profile_edit)
-    }
-}
