+++ app/src/main/res/layout/fragment_search.xml
@@ -33,7 +33,7 @@
         <androidx.appcompat.widget.SearchView
             android:id="@+id/searchView"
             android:layout_width="match_parent"
-            android:layout_height="wrap_content"
+            android:layout_height="40dp"
             app:queryHint="Search here..."
             app:iconifiedByDefault="false"
             app:queryBackground="@color/transparent" />
