+++ app/src/main/res/layout/fragment_setting.xml
@@ -48,7 +48,7 @@
                 android:layout_height="wrap_content"
                 android:layout_marginTop="12dp"
                 android:text="@string/setting_test_id"
-                android:textSize="13sp"
+                android:textSize="13dp"
                 android:textColor="@color/black"
                 app:layout_constraintStart_toStartOf="parent"
                 app:layout_constraintEnd_toEndOf="parent"
@@ -61,7 +61,7 @@
                 android:layout_marginTop="12dp"
                 android:textColor="@color/black"
                 android:text="@string/setting_profile_modify"
-                android:textSize="13sp"
+                android:textSize="13dp"
                 app:layout_constraintStart_toStartOf="parent"
                 app:layout_constraintEnd_toEndOf="parent"
                 app:layout_constraintTop_toBottomOf="@id/tvId"/>
@@ -72,7 +72,7 @@
                 android:layout_height="wrap_content"
                 android:layout_marginTop="24dp"
                 android:text="@string/setting_participated_challenge"
-                android:textSize="12sp"
+                android:textSize="12dp"
                 android:textColor="@color/black"
                 app:layout_constraintStart_toStartOf="parent"
                 app:layout_constraintTop_toBottomOf="@id/btnProfileModify"/>
@@ -97,7 +97,7 @@
                 android:layout_marginTop="28dp"
                 android:textColor="@color/black"
                 android:text="@string/setting_logout"
-                android:textSize="15sp"
+                android:textSize="15dp"
                 app:layout_constraintStart_toStartOf="parent"
                 app:layout_constraintEnd_toEndOf="parent"
                 app:layout_constraintTop_toBottomOf="@id/rvChallenge"/>
@@ -109,7 +109,7 @@
                 android:layout_marginTop="12dp"
                 android:textColor="@color/black"
                 android:text="@string/setting_secession"
-                android:textSize="15sp"
+                android:textSize="15dp"
                 android:layout_marginBottom="24dp"
                 app:layout_constraintStart_toStartOf="parent"
                 app:layout_constraintEnd_toEndOf="parent"
