+++ app/src/main/res/layout/item_participated_challenge.xml
@@ -16,7 +16,7 @@
         android:layout_marginTop="16dp"
         android:text="@string/setting_hashtag"
         android:textColor="@color/black"
-        android:textSize="13sp"
+        android:textSize="13dp"
         android:textStyle="bold"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent" />
@@ -28,7 +28,7 @@
         android:layout_marginStart="4dp"
         android:text="@string/setting_challenge_title"
         android:textColor="@color/black"
-        android:textSize="13sp"
+        android:textSize="13dp"
         android:textStyle="bold"
         app:layout_constraintBottom_toBottomOf="@id/tvHashtag"
         app:layout_constraintStart_toEndOf="@id/tvHashtag"
@@ -41,7 +41,7 @@
         android:layout_marginTop="8dp"
         android:text="@string/setting_challenge_content"
         android:textColor="#CBCBCB"
-        android:textSize="13sp"
+        android:textSize="13dp"
         app:layout_constraintStart_toStartOf="@id/tvHashtag"
         app:layout_constraintTop_toBottomOf="@id/tvHashtag" />
 
@@ -61,7 +61,7 @@
         android:layout_height="wrap_content"
         android:text="@string/setting_challenge_participant"
         android:textColor="@color/black"
-        android:textSize="10sp"
+        android:textSize="10dp"
         android:layout_marginEnd="20dp"
         app:layout_constraintTop_toTopOf="@id/ivProfile"
         app:layout_constraintBottom_toBottomOf="@id/ivProfile"
@@ -73,7 +73,7 @@
         android:layout_height="wrap_content"
         android:text="@string/setting_Dday"
         android:textColor="@color/black"
-        android:textSize="13sp"
+        android:textSize="13dp"
         android:background="@drawable/bg_dday"
         android:gravity="center"
         android:layout_marginTop="8dp"
